refactor code into parser.util
