+++ src/main/java/seedu/zerotoone/logic/parser/util/Prefix.java
@@ -1,4 +1,4 @@
-package seedu.zerotoone.logic.parser;
+package seedu.zerotoone.logic.parser.util;
 
 /**
  * A prefix that marks the beginning of an argument in an arguments string.
